requirements.txt
leaflet
django_google_maps
mapwidget
# crispy_forms

